quantidade e consumo alterado para float, alterado html e css, inserido botao para sair, bloqueado rotas onlyUsers

+++ src/models/Consumo.js
@@ -3,7 +3,7 @@ const { Model, DataTypes } = require('sequelize')
 class Consumo extends Model {
     static init(sequelize) {
         super.init({
-            quantidade_consumida: DataTypes.INTEGER,
+            quantidade_consumida: DataTypes.FLOAT,
         }, {
             sequelize
         })
